removed some references to longValue for uint64
DepositTree is probably the class i had the biggest go at, most are just trivials.

+++ validator/client/src/main/java/tech/pegasys/teku/validator/client/doppelganger/DoppelgangerDetector.java
@@ -215,7 +215,7 @@ public class DoppelgangerDetector {
               .filter(pubKey -> !detectedDoppelgangers.containsValue(pubKey))
               .collect(Collectors.toSet());
       statusLog.doppelgangerCheck(
-          epoch.longValue(), mapToAbbreviatedKeys(inactivePubKeys).collect(Collectors.toSet()));
+          epoch, mapToAbbreviatedKeys(inactivePubKeys).collect(Collectors.toSet()));
 
       return validatorApiChannel
           .getValidatorIndices(inactivePubKeys)
